move lambda layer to asset folder
