--- scripts/layers/weather/requirements.txt
@@ -1,1 +1,0 @@
-meteostat
