+++ source/lambda/feature_volt_var_input/requirements.txt
@@ -1,0 +1,1 @@
+pyathena
