+++ source/lambda/topology_transformer/requirements.txt
@@ -1,0 +1,1 @@
+fastparquet
