+++ templates/feature/voltvar/voltvar.feature.template.yaml
@@ -325,23 +325,6 @@ Resources:
       Timeout: 240
       ReservedConcurrentExecutions: 1
       Handler: 'app.lambda_handler'
-      Layers:
-        - !Ref CrhelperLayer
       Code:
         S3Bucket: !If [ UsingDefaultBucket, !Sub '${QSS3BucketName}-${AWS::Region}', !Ref QSS3BucketName ]
         S3Key: !Sub '${QSS3KeyPrefix}assets/lambda/s3_copy_custom_resource/lambda.zip'
-
-  CrhelperLayer:
-    Type: AWS::Lambda::LayerVersion
-    Properties:
-      LayerName: 'Crhelper'
-      Description: 'Crhelper'
-      CompatibleRuntimes:
-        - 'python3.9'
-      CompatibleArchitectures:
-        - 'x86_64'
-      Content:
-        S3Bucket: !If [ UsingDefaultBucket, !Sub '${QSS3BucketName}-${AWS::Region}', !Ref QSS3BucketName ]
-        S3Key:
-          Fn::Sub: '${QSS3KeyPrefix}scripts/layers/crhelper.zip'
-      LicenseInfo: 'Apache 2.0'
